requirements.txt
asgiref==3.8.1
bleach==6.2.0
Django==5.1.2
django-jazzmin==3.0.1
django-restframework==0.0.1
djangorestframework==3.15.2
drf-api-logger==1.1.16
modeltranslation==0.25
sqlparse==0.5.1
tzdata==2024.2
webencodings==0.5.1
